refactor(checker): isolate exact rewrite query boundary
